feat(topBar): Refactor assets and components structure
