--- client/src/components/button/index.tsx
@@ -1,2 +1,0 @@
-import {buttonVariants} from "@/components/ui/button"
-
