+++ client/src/components/index.tsx
@@ -1,0 +1,5 @@
+import topBar from "./topBar";
+
+export {
+     topBar as TopBar
+    };
