+++ client/src/components/topBar/index.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { LeftTopBar, RightTopBar } from "../../assets";
+import { LeftTopBar, RightTopBar } from "@/assets";
 
 type Props = {
   page1: React.ReactNode;
@@ -13,19 +13,36 @@ export default function topBar({page1, page2}: Props) {
   return(
 
     <div>
-      
-
-      <Tabs defaultValue="service" className="flex flex-col items-center bg-white">
+      <Tabs 
+        defaultValue="service" 
+        className="flex flex-col items-center bg-white"
+      >
         
-        <TabsList className="pt-5 pb-5 bg-white text-[#242424] flex flex-1 flex-row justify-around h-[114px] border-b-2 rounded-none border-solid border-[#D9D9D9] w-full items-center mb-5">
-          <Image src={LeftTopBar} alt="Logo CITi Vet"/>
+        <TabsList className="py-5 px-12 bg-white text-[#242424] flex flex-1 flex-row justify-between h-[114px] border-b-2 rounded-none border-solid border-[#D9D9D9] w-full items-center mb-5">
+          <Image 
+            src={LeftTopBar} 
+            alt="Logo CITiVet"
+          />
 
-          <div className="pb-2 mt-6 ">
-            <TabsTrigger value="service" className="font-normal p-0 pb-1 mr-6 mb-0 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-[#50E678]">Atendimento</TabsTrigger>
-            <TabsTrigger value="register" className="font-normal p-0 pb-1 ml-6 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-[#50E678]">Cadastro</TabsTrigger>
+          <div className="pb-2 mt-6">
+            <TabsTrigger 
+              value="service" 
+              className="font-normal p-0 pb-1 mr-6 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-[#50E678]"
+            >
+              Atendimento
+            </TabsTrigger>
+            <TabsTrigger 
+              value="register" 
+              className="font-normal p-0 pb-1 ml-6 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-[#50E678]"
+            >
+              Cadastro
+            </TabsTrigger>
           </div>
 
-          <Image src={RightTopBar} alt="CITi Slogan"/> 
+          <Image 
+            src={RightTopBar} 
+            alt="CITi Slogan"
+          /> 
         </TabsList>
 
         <TabsContent value="service">
@@ -39,5 +56,5 @@ export default function topBar({page1, page2}: Props) {
       </Tabs>
     </div>
 
-  )
+  );
 }
